% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov_CR3J}
\alias{vcov_CR3J}
\title{Compute CR3 Jackknive variance covariance matrices of objects of type lm and fixest}
\usage{
vcov_CR3J(object, ...)
}
\arguments{
\item{object}{An object of class \code{lm} or `fixest``}

\item{...}{Other arguments}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects of type lm and fixest
}
