% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcovJN}
\alias{vcovJN}
\title{Implements the fast algorithm for CRV3 Jackknife estimation in MacKinnon, Nielsen & Webb (2022)}
\usage{
vcovJN(model, clustid, type = "CRV3", ...)
}
\arguments{
\item{model}{An object of type lm}

\item{clustid}{A clustering vector}

\item{type}{'CRV3' or 'CRVJ3'}

\item{...}{other function arguments passed to 'vcov'}
}
\description{
Implements the fast algorithm for CRV3 Jackknife estimation in MacKinnon, Nielsen & Webb (2022)
}
