% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov_CR3J.fixest}
\alias{vcov_CR3J.fixest}
\title{Compute CR3 Jackknive variance covariance matrices of objects of type fixest}
\usage{
\method{vcov_CR3J}{fixest}(obj, cluster, ...)
}
\arguments{
\item{cluster}{A clustering vector}

\item{...}{other function arguments passed to 'vcov'}

\item{object}{An object of type lm}
}
\description{
Compute CR3 Jackknive variance covariance matrices of objects of type fixest
}
